Region tooltips

+++ src/main/kotlin/app/utils/Math.kt
@@ -2,8 +2,8 @@ package app.utils
 
 import kotlin.math.pow
 
-fun sq(x: Int): Int {
-    return x * x
+fun sq(x: Int): Double {
+    return (x * x).toDouble()
 }
 
 fun sq(x: Double): Double {
